Rename RetryError to RetryErrorType
